Add files via upload
This Frontend Mentor challenge (https://www.frontendmentor.io/challenges/interactive-pricing-component-t0m8PIyY8) contains HTML5, CSS3 and vanilla JS.
It took me a day to complete, including the JS and the responsiveness. 
It is listed as a Junio Level challenge (level 2). It was quite easy for me, but I still learned a few things and refreshed my memory. 
I learned Flexbox through this challenge.
I am aware that the design is not 100% the same as the one in the challenge. There are some minor differences: font size, padding, etc. I didn't follow the style guide to the letter because I got carried away with the coding. 
All-in-all, I enjoyed coding it so much, and I'm proud of how it turned out.